I put do and printInfo in one file called tools, for I use them in both IPA and APK
