+++ cmd/tools.go
@@ -2,9 +2,25 @@ package cmd
 
 import (
 	"fmt"
+	"os/exec"
 
 	"github.com/bitrise-io/go-utils/log"
 )
+
+func do(c command) error {
+	if c.extraFlag != "" {
+		_, err := exec.Command(c.Command, c.Flag, c.Path, c.extraFlag, c.outPath).Output()
+		if err != nil {
+			return err
+		}
+	} else {
+		_, err := exec.Command(c.Command, c.Flag, c.Path).Output()
+		if err != nil {
+			return err
+		}
+	}
+	return nil
+}
 
 func printAppInfo(appInfo map[string]string, appType string) {
 	if appType == "apk" {
@@ -20,4 +36,12 @@ func printAppInfo(appInfo map[string]string, appType string) {
 		log.Printf("Build Number: %s", appInfo["Build Number"])
 		log.Printf("Path to icon: %s", appInfo["App Icon"])
 	}
+}
+
+type command struct {
+	Command   string
+	Flag      string
+	Path      string
+	extraFlag string
+	outPath   string
 }
